Replace Users class to UsrMgmt class

+++ app/apis/totp.py
@@ -1,8 +1,8 @@
 """Blueprint API."""
+from flask import current_app as ca
 from flask_restx import Namespace, Resource, abort
 
 from ..helpers.decorator import role_required, token_required
-from ..helpers.users import User, UserNotFound
 from .models import message, otp
 
 api = Namespace(
@@ -26,47 +26,39 @@ class Totp(Resource):
     @api.response(422, "Error", message)
     def get(self, id: int):  # pylint: disable=W0622
         """Get OTP for a user."""
-        try:
-            user = User(id)
+        if user := ca.usrmgmt.get(id=id):
             user.set_secret()
             return user
-        except UserNotFound:
-            abort(404, "User not found")
+        abort(404, "User not found")
 
     @api.response(204, "Action is success")
     @api.response(404, "Not found", message)
     def post(self, id: int):  # pylint: disable=W0622
         """Check OTP code."""
-        try:
-            user = User(id)
+        if user := ca.usrmgmt.get(id=id):
             if user.totp is True:
                 if user.check_totp(api.payload["secret"]):
                     return "", 204
                 abort(422, "OTP incorrect")
-            abort(422, "OTP not enable")
-        except UserNotFound:
-            abort(404, "User not found")
+            abort(422, "OTP not enabled")
+        abort(404, "User not found")
 
     @api.response(204, "Action is success")
     @api.response(404, "Not found", message)
     @api.response(422, "Error", message)
     def put(self, id: int):  # pylint: disable=W0622
         """Check and create OTP Code for a user."""
-        try:
-            user = User(id)
+        if user := ca.usrmgmt.get(id=id):
             if user.validate_secret(api.payload["secret"]):
                 return "", 204
             abort(422, "OTP incorrect")
-        except UserNotFound:
-            abort(404, "User not found")
+        abort(404, "User not found")
 
     @api.response(204, "Action is success")
     @api.response(404, "Not found", message)
     def delete(self, id: int):  # pylint: disable=W0622
         """Delete OTP infos for a user."""
-        try:
-            user = User(id)
+        if user := ca.usrmgmt.get(id=id):
             user.delete_secret()
             return "", 204
-        except UserNotFound:
-            abort(404, "User not found")
+        abort(404, "User not found")
